+++ ControlStochasticAge.cs
@@ -10,6 +10,12 @@ using System.Windows.Forms;
 
 namespace Nmfs.Agepro.Gui
 {
+    public enum StochasticAgeFleetDependency
+    {
+        dependent,
+        independent
+    };
+
     public partial class ControlStochasticAge : UserControl
     {
         protected ControlStochasticAgeDataGridTable controlStochasticParamAgeFromUser;
@@ -25,6 +31,9 @@ namespace Nmfs.Agepro.Gui
             radioParameterFromUser.Checked = true; //User Specfied Option Selected by Default
             stochasticParameterLabel = "Stochastic Parameters"; //Default Fallback Text
             settingParameterForControl = false;
+
+            //By Default, Stochastic Parameters are fleet independent.
+            fleetDependency = StochasticAgeFleetDependency.independent;
 
             controlStochasticParamAgeFromFile.timeVaryingFileChecked += 
                 new EventHandler(linkTimeVaryingUserSpecAndFromFile);
@@ -77,6 +86,12 @@ namespace Nmfs.Agepro.Gui
             get { return controlStochasticParamAgeFromUser.enableTimeVaryingCheckBox; }
             set { controlStochasticParamAgeFromUser.enableTimeVaryingCheckBox = value; }
         }
+        public StochasticAgeFleetDependency fleetDependency 
+        {
+            get { return controlStochasticParamAgeFromUser.fleetDependent; }
+            set { controlStochasticParamAgeFromUser.fleetDependent = value; }
+        }
+
 
         protected void linkTimeVaryingUserSpecAndFromFile(object sender, EventArgs e)
         {
@@ -109,6 +124,33 @@ namespace Nmfs.Agepro.Gui
 
             base.OnLoad(e);
         }
+
+        /// <summary>
+        /// Generalized method to load Stochastic Data Parameters from AGEPRO Input Data files.
+        /// </summary>
+        /// <param name="inp">AGEPRO InputFile StochasticAge Parameters </param>
+        /// <param name="generalOpt">AGEPRO InputFile General Options values</param>
+        public void LoadStochasticAgeInputData(Nmfs.Agepro.CoreLib.AgeproStochasticAgeTable inp,
+            Nmfs.Agepro.CoreLib.AgeproGeneral generalOpt)
+        {
+            this.readInputFileState = true;
+            this.seqYears = Array.ConvertAll(generalOpt.SeqYears(), element => element.ToString());
+            this.numFleets = generalOpt.numFleets;
+            this.timeVarying = inp.timeVarying;
+            this.stochasticDataFile = inp.dataFile;
+            this.stochasticAgeTable = Util.GetAgeproInputDataTable(this.stochasticAgeTable, inp.byAgeData);
+            this.stochasticCV = Util.GetAgeproInputDataTable(this.stochasticCV, inp.byAgeCV);
+            if (!(this.stochasticAgeTable != null))
+            {
+                this.enableTimeVaryingCheckBox = false;
+            }
+            else
+            {
+                this.enableTimeVaryingCheckBox = true;
+            }
+            this.readInputFileState = false;
+        }
+
         protected virtual void radioParameterFromUser_CheckedChanged(object sender, EventArgs e)
         {
             if (settingParameterForControl == false)
@@ -130,6 +172,80 @@ namespace Nmfs.Agepro.Gui
             }
             
         }
+
+        /// <summary>
+        /// Creates a empty Data Table for the Stochastic Parameter Control based on the user inputs gathered 
+        /// from the General Options control parameter.
+        /// </summary>
+        /// <param name="ctl">Stochastic Parameter Control</param>
+        /// <param name="genOpt">Paramters from the General Options Control</param>
+        /// <param name="fleetDependent">Is this Stochastic Parameter dependent on the 
+        /// nubmber of fleets? Default is false.</param>
+        public void CreateStochasticParameterFallbackDataTable(ControlGeneral genOpt,
+             StochasticAgeFleetDependency fleetDependent = StochasticAgeFleetDependency.independent)
+        {
+            this.numFleets = Convert.ToInt32(genOpt.generalNumberFleets);
+            this.seqYears = genOpt.SeqYears();
+            this.readInputFileState = true;
+            //Reset Tables if they were used before
+            if (this.stochasticAgeTable != null)
+            {
+                this.stochasticAgeTable.Reset();
+                this.stochasticCV.Reset();
+            }
+
+            if (this.timeVarying == true)
+            {
+                this.stochasticAgeTable = CreateFallbackAgeDataTable(genOpt.NumAges(),
+                    genOpt.SeqYears().Count(), this.numFleets);
+            }
+            else
+            {
+                if (fleetDependent == StochasticAgeFleetDependency.dependent)
+                {
+                    this.stochasticAgeTable = CreateFallbackAgeDataTable(genOpt.NumAges(), 1, this.numFleets);
+                    this.stochasticCV = CreateFallbackAgeDataTable(genOpt.NumAges(), 1, this.numFleets);
+                }
+                else
+                {
+                    this.stochasticAgeTable = CreateFallbackAgeDataTable(genOpt.NumAges(), 1, 1);
+                    this.stochasticCV = CreateFallbackAgeDataTable(genOpt.NumAges(), 1, 1);
+                }
+
+            }
+            
+            this.readInputFileState = false;
+
+        }
+
+        /// <summary>
+        /// Creates an empty DataTable for AGEPRO Parameter Control based on the number ages and years (or 
+        /// fleet-years).
+        /// </summary>
+        /// <param name="numAges">Number of Age Classes.</param>
+        /// <param name="numYears">Number of Years (from First year to Last Year of projection)</param>
+        /// <param name="numFleets">Number of Fleets. Default is 1</param>
+        /// <returns>Returns a empty DataTable</returns>
+        public DataTable CreateFallbackAgeDataTable(int numAges, int numYears, int numFleets = 1)
+        {
+            int numFleetYears = numYears * numFleets;
+
+            DataTable fallbackTable = new DataTable();
+
+            for (int icol = 0; icol < numAges; icol++)
+            {
+                fallbackTable.Columns.Add("Age" + " " + (icol + 1));
+            }
+            for (int row = 0; row < numFleetYears; row++)
+            {
+                fallbackTable.Rows.Add();
+            }
+            Nmfs.Agepro.CoreLib.Extensions.FillDBNullCellsWithZero(fallbackTable);
+
+            return fallbackTable;
+        }
+
+
 
         public virtual bool ValidateStochasticParameter(int numAges)
         {